Fixing memory management with params
I did this crazy thing long ago where every other string in the array way
const, and you had to use a custom free function.  And somewhere along the
way, the custom free function got lost, and it rotted into something that
leaked half its memory and freed the other half twice.  So we're just going
to go ahead and strdup and stop with the premature optimization.