Use GString instead of a plain string to hold content
Yelp was holding the string that it gives WebKit for loading in a
plain string, and ignoring the explicit length argument. In some cases
the string was not correctly terminated by \0, and the document ended
up with garbage. This was causing most of the XML parsing error
messages that were being observed.

This patch was made with help from Luciana Fujii <luciana@fujii.eti.br>.