Fixing issues introduced by xulrunner 1.9.2
API changes in a stable release are not fun.  With 1.9.2, setting the base
URI when opening a stream causes a uri_selected callback.  So we compare
the URI to what we know we're loading to prevent an infinite loop.

Also, #fragment links no longer trigger a uri_selected callback.  So I've
halfway backported the xref: method used in the yelp-3-0 branch.  It's not
pervasive.  We rewrite it pretty much immediately.  So we still use the
slash hack, instead of the nicer method in yelp-3-0.  This should be enough
to keep 2.x alive.