[yelp-window.c] Temporarily put frag IDs in location entry
We're getting confusing seemingly duplicate entries in the location
entry when we go to fragments of pages.  We want to get the actual
fragment title, possibly mixed with the page title.  That requires
some work on YelpDocument and YelpView.  For now, put in the frag
ID to disambiguate and remind me to fix this properly.