[yelp-uri] Use '#' to separate page and frag in file URIs
This is all internal bookkeeping. When we have to put both a page ID
and a fragment identifier on a file URI, we previously separated them
with a slash. But that interfered with the way we used leading slashes
for special pages. So we just use a second hash instead.