Allow the possibility of no compression for info files.
At the moment, the info parser uses YelpMagicDecompressor. If it
didn't recognize a stream as BZ2 or LZMA, the decompressor assumed it
was GZip. This didn't work if the file is uncompressed... Since GZip
decompression certainly won't work if the file is not GZip compressed,
this patch changes YelpMagicDecompressor to fall back to a
pass-through converter if it doesn't recognise the compression format
(which might mean the file is uncompressed!).