[libyelp] Show bookmarks in location entry with YelpBookmarks
The bookmarks are managed by the application, not libyelp, so we need a way for
the application to tell libyelp about bookmarks. Add a YelpBookmarks interface,
which is implemented by YelpApplication. Then we can pass a YelpBookmarks to
any libyelp class that needs it.

Still need to hook up the bookmarks-changed signal and do something useful when
the bookmark-new icon is clicked.