Deal correctly with the last line of output.
Since we're parsing the whole thing in one pass, we can't know this is
the last line until we're done, so we go through at the end and check
whether there's an empty section. If so, we know what the title should
look like, so we extract the information and put it somewhere more
useful.