Remove bookmark directory from page, no editor, use symbolic
I've removed the bookmarks editor dialog, because you can't do
much editing anyway. Instead, you can now remove a bookmark to
the page you're on. There's a menu item for this, and the icon
in the location entry is now a toggle.

I did the toggle icon with the new symbolic icons, changing
the fill color for bookmarked/unbookmarked. And while I was
in there, I used symbolic icons for the search and clear
icons in the search entry as well.