Make the man: uri parser deal with more man page names
The uri parser used to support both man:printf(1) and man:printf.1,
but the support for the . variant comes at the cost of not supporting
man:printers.conf. Since embedded periods in man pages are not that
uncommon, in particular for file format documentation, this patch
allows periods inside the name part of a man: uri, as long as it
is not followed by a digit. This should cover pretty much all naturally
occurring embedded periods, while still supporting the man:printf.1
syntax.

https://bugzilla.gnome.org/show_bug.cgi?id=676482