yelp-view: Implement pages load in WebKit2
Substitute webkit_web_view_load_string call for webkit_web_view_load_html for showing error pages.
Implement custom uri schemes for loading the normal pages.

We should deal with the uris that Yelp undestand and the uris that Webkit undestand so we have created functions to convert from one to the others. In addition, we should add a hack to be able to load absolute uris. When we have a help:gnome-help/... uri on our document and the current page has the same scheme (help) WebKit interprets this uri as relative and it builds a different uri. To fix this instead of use a help scheme we use a bogus-help schme so WebKit interprets the uri as absolute.