Use GtkSettings 'gtk-xft-dpi' property to keep track of Xft DPI changes
Use that property to find the DPI resolution for font handling, in order
to calculate the effective font size value that needs to be passed to
WebKitSettings when setting the default values. Also, connect to
notify::gtk-xft-dpi so that we can update the font size whenever the
DPI value changes, and not just on startup.

https://bugzilla.gnome.org/show_bug.cgi?id=744921