Avoid activating ourselves recursively
We consider INFO:foobar an 'external uri', and call
g_app_info_launch_default_for_uri on it. Which is bad,
since GIO promptly determines the default handler for INFO:
uris to be...yelp. Avoid this by manually looking up the
default handler, and ignoring it if it appears to be yelp.

https://bugzilla.gnome.org/show_bug.cgi?id=745407