web-extension: Fix external resources of ghelp documents
ghelp uris can start with the document name or with a slash and point
directly to a xternal resource like an image. However, we were assuming the
uris always started with the document name, so for uris not including
it, we were removing the first element of the path, typically the
images/figures directory. We could check first if the uri starts with
the document uri, to remove the document uri prefix from the path or
only the scheme.