yelp-uri-builder: Handle info uris
info uris were not handled by build_network_uri and build_yelp_uri. We
need to make sure they are valid network URIs from the WebKit point of
view, like we do for other URIs, but we also need to handle the fragment
part differently. When navigating from info:bar to info:foo#bar, WEbKoit
doesn't start a new load, since it's considered a navigation inside the
same already loaded page. So, we need to make WebKit think this is a new
path, so that the load happens and the new section is resolved and
loaded. To do this we just replace the '#' by '/' in build_network_uri
and the '/' by '#' in build_yelp_uri.