yelp-view: Update current page information when loaded from history cache
When a page is loaded from the history cache, the URI is not resolved
and the metadata is not updated. Check in policy callback if the action
is a back-forward navigation and resolve the URI in that case but
without loading the page once the URI is resolved, since the page will
be loaded from the history cache.

https://bugzilla.gnome.org/show_bug.cgi?id=753448