Fallback to 'help-list:' if the default URI is not available
The current default URI when none is explicitly specified is "help:gnome-help",
which maps to the GNOME 3 desktop users guide. If this is not specified, yelp
now will fallback to showing 'help-list:', which is always available.

The new fallback applies both when launched through command line or when a new
yelp window is requested.

The patch also avoids a YelpUri memleak that was happening in
yelp_application_new_window(), as yelp_application_new_window_uri() was itself
taking its own reference.

https://bugzilla.gnome.org/show_bug.cgi?id=759545