Simplify desktop file generation
Desktop file was processed twice - by the configure script and by
intltool. Configure was putting a yelp version to some
bugzilla-related value. Bugzilla is an obsolete bug tracker in GNOME
(the project has moved to GNOME gitlab), so the field is also
obsolete. With dropping the fields, configure has no business in the
desktop file any more.