Move from intltool to gettext
This is to make a migration from autotools to meson easier. Also,
intltool got obsoleted by gettext.

There are minor differences in some of the translated files, but
functionally it should not make any difference. But why gettext insist
on reversing the alphabetic order in the appdata file (so zh_TW comes
first and cs comes last) is beyond me.

https://gitlab.gnome.org/GNOME/yelp/merge_requests/11