Reorganize icons a bit
Most of the icons go to the /usr/share/yelp/icons directory, and only
two other icons go to /usr/share/icons. Separate both sets of icons,
so the former icons are in the yelp-icons directory and the latter are
in icons directory. This will make installation easier on meson.