Be more careful about stripping slashes
The transition to GUri introduced some problems with our URI mangling,
because it's more strict about URI syntax than libsoup was. This led
to absolute file paths being broken. This ought to do it. Some day I
might redo all of this, because it's a sloppy pile of legacy that's
been monkeypatched thru half a dozen API changes in other libraries.
That day is not today.

Fixes #182
cf https://bugzilla.redhat.com/show_bug.cgi?id=1997839