yelp-window: switch to .ui file
Drop the window creation code in favor of creating it through a .ui
file. The only things that stay in the yelp-window.c file are some
signal bindings (they pass the window as the user data), creating
the search entry (since we need to pass the view and a Bookmarks
object to it), popover assignment to the bookmark menu button and
the drop target creation.

The CSS style that was set in the code is also moved to a separate
file, style.css, and is included as part of the GResource.