meta: update dependencies to GTK4, libadwaita and WebKitGTK 6.0
GTK+ 3.0 is replaced by GTK 4.0, libhandy is replaced by libadwaita
and WebKit2GTK 4.x is replaced by WebKitGTK 6.0.

GTK 4.16.0 and libadwaita 1.6.0 are set as the minimum required
versions, as they are currently the latest available versions.
These requirements will inform later porting decisions.

Nothing builds with this combination yet; this is just the commit
that makes the build systems, Autotools and Meson, aware that we
are building with GTK4 now.