yelp-search-entry: switch from entry completion API to listview, redesign
Switch away from the now deprecated entry completion APIs and create a
custom search popover which uses a ListView internally.

This brings with it a couple of internal changes:

- The search entry gains its own UI file.

- A new object type, YelpSearchResult, is introduced to contain information
  about the individual search results. Previously it was stored in a 6-item
  GtkListStore; while we *could* try to use such a structure here, it was
  much easier to turn it into an object so that we could benefit from the
  property lookup functionality in the listview UI factory.

- YelpBookmarks gains two new signals: bookmark-added and bookmark-removed.
  This is done to facilitate updating bookmarked status for the search entries.

The design of the search results has been completely reworked. The new
design is taken directly from Web/Epiphany; the UI file was derived from
Epiphany's location entry code. Some changes were made; in particular,
the search results are now limited to only 7 entries (6 results and 1 space
for the "Search for ..." button). This is to make the full-blown search more
visible, but it can be easily reverted if necessary (Epiphany actually shows
more entries and allows the user to scroll through them, but I thought this
felt janky with Yelp's imprecise matching...)

A potential improvement would be to highlight the matched part in bold,
but I'll leave that for later :)